requirements.txt
Flask>=2.3.0
Werkzeug>=2.3.0
PyPDF2>=3.0.0
reportlab>=4.0.0
Pillow>=9.0.0
PyMuPDF>=1.23.0
pathlib2>=2.3.0
